Reference activeSource with a useRef object
The event handler function defined in the one-time useEffect setup
function is bound to the state that was present during the render it was
created in. We only want to create this hook once, but see state from
the most recent render. This is accomplished via useRef, which gives
access to a mutable object.

https://reactjs.org/docs/hooks-faq.html#why-am-i-seeing-stale-props-or-state-inside-my-function