Pass placeholder Level struct into Home and Flag templates
Go templating evaluates all arguments to a function. As a result, and/or
operators will fail under certain conditions when Level is a string rather than a Level struct. Having `Level` potentially be a non Level type is probably bad Go practice anyway. Keeping this type consistent somewhat simplifies the logic of the nav.

This issue with operator evaluation is described in depth here, and
appears to be something that will change:

https://github.com/golang/go/issues/31103