Use inotify to monitor for a reset signal
This is a way for one container to send a signal to another
without mounting the docker socket into the former. This then allows the
latter to reset it's state, which is useful because some solutions put
containers into an unstable state.

inotify watches a directory for changes, and resets application state
and the process when the reset pattern matches.