Added a cloud server for level5
Level 5 is special in that it communicates directly with level2, but
from the server rather than from a browser. Including the token
generation in the server code might give too much away, so token
generation has been moved into a server subclass, which will be run in
the cloud in place of the regular server.